refactor: extract PlaybookEditor component from MainLayout
PlaybookEditor (299 lines) encapsulates all editor logic: playbook loading,
collaboration WebSocket connect/disconnect, sync callbacks, WorkZone,
ConfigZone, SystemZone with resize. MainLayout (230 lines) keeps only
AppHeader, left panel (Project/Modules tabs), and PlaybookManagerDialog.
